+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/batch/TransactionalBulkExecutor.java
@@ -44,6 +44,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Supplier;
 
 import static com.azure.cosmos.implementation.batch.BatchRequestResponseConstants.DEFAULT_MAX_TRANSACTIONAL_BATCH_INTERVAL_AFTER_DRAINING_INCOMING_FLUX_IN_MILLISECONDS;
 import static com.azure.cosmos.implementation.guava25.base.Preconditions.checkNotNull;
@@ -830,6 +831,7 @@ public final class TransactionalBulkExecutor implements Disposable {
             docClientWrapper,
             container,
             cosmosBatchBulkOperation.getPartitionKeyValue(),
+            (Supplier<String>) null,
             null);
     }